+++ extensions/vscode-create-rax/src/templates/create.html.ejs
